+++ drivers/sqljs-driver/src/jsTest/kotlin/com/squareup/sqldelight/drivers/sqljs/JsDriverTest.kt
@@ -28,7 +28,7 @@ class JsDriverTest {
   }
 
   private val schema = object : SqlSchema<QueryResult.Value<Unit>> {
-    override val version: Int = 1
+    override val version: Long = 1
 
     override fun create(driver: SqlDriver): QueryResult.Value<Unit> {
       driver.execute(
@@ -59,8 +59,8 @@ class JsDriverTest {
 
     override fun migrate(
       driver: SqlDriver,
-      oldVersion: Int,
-      newVersion: Int,
+      oldVersion: Long,
+      newVersion: Long,
       vararg callbacks: AfterVersion,
     ): QueryResult.Value<Unit> {
       // No-op.
